feat: 엔진, 실행결과 항목 변경
