--- src/app/result/solution-log/page.js
@@ -1,9 +1,0 @@
-"use client";
-
-export default function ToolUsagePage() {
-  return (
-    <div style={{ padding: "20px", fontSize: "16px", color: "#555" }}>
-      📌 이 페이지에는 용액 일지 조회 내용 기입 예정
-    </div>
-  );
-}
