--- src/app/result/tool-usage/page.js
@@ -1,9 +1,0 @@
-"use client";
-
-export default function ToolUsagePage() {
-  return (
-    <div style={{ padding: "20px", fontSize: "16px", color: "#555" }}>
-      📌 이 페이지에는 작업도구 사용 내역 내용 기입 예정
-    </div>
-  );
-}
