+++ src/app/result/workcenter-plan/page.js
@@ -7,38 +7,52 @@ import ResultToolBar from "@/app/standard/ResultToolBar";
 
 const columns = [
   { field: "id", headerName: "순번", width: 80 },
-  { field: "workcenter", headerName: "작업장", width: 150 },
-  { field: "planDate", headerName: "계획 일자", width: 130 },
-  { field: "productName", headerName: "제품명", width: 180 },
-  { field: "planQty", headerName: "계획 수량", flex:1 },
+  { field: "workcenterId", headerName: "작업장 ID", width: 120 },
+  { field: "workcenterName", headerName: "작업장 명", width: 150 },
+  { field: "operationId", headerName: "공정 ID", width: 130 },
+  { field: "operationName", headerName: "공정명", width: 150 },
+  { field: "operationType", headerName: "공정 타입", width: 130 },
+  { field: "workcenterStartTime", headerName: "시작 시간", width: 180 },
+  { field: "workcenterEndTime", headerName: "종료 시간", width: 180 },
+  { field: "workcenterGroup", headerName: "작업장 그룹", width: 130 },
+  { field: "toolId", headerName: "설비 ID", width: 100 },
+  { field: "toolName", headerName: "설비명", width: 100 },
+  { field: "routingId", headerName: "라우팅 ID", width: 130 },
 ];
 
 export default function WorkcenterPlanView() {
   const [rows, setRows] = useState([]);
+  const [total, setTotal] = useState(0);
 
   useEffect(() => {
-    fetch("http://localhost:8080/api/result/workcenter-plan")
+    const scenarioId = "S000001"; // 시나리오 ID 고정
+    fetch(`http://localhost:8080/api/result/workcenter-plan/${scenarioId}`)
       .then((res) => res.json())
       .then((data) => {
-        const formatted = (data?.resultList || []).map((item, index) => ({
-          id: index + 1,
-          workcenter: item.workcenter,
-          planDate: item.planDate ?? "-",
-          productName: item.productName ?? "-",
-          planQty: item.planQty ?? 0,
+        const plans = data.plans || [];
+        const formatted = plans.map((plan) => ({
+          id: plan.no, // DataGrid 고유 ID
+          ...plan,
+          workcenterStartTime:
+            plan.workcenterStartTime?.replace("T", " ") ?? "-",
+          workcenterEndTime: plan.workcenterEndTime?.replace("T", " ") ?? "-",
         }));
         setRows(formatted);
+        setTotal(data.total || 0);
+      })
+      .catch((err) => {
+        console.error("데이터 불러오기 실패 ❌", err);
       });
   }, []);
 
   return (
     <Box sx={{ display: "flex", flexDirection: "column", height: "100%" }}>
-      {/* 툴바 영역 (업로드/다운로드 버튼 자리) */}
+      {/* 툴바 영역 */}
       <Box sx={{ mt: 2, mb: 1 }}>
         <ResultToolBar upload={() => {}} download={() => {}} />
       </Box>
 
-      {/* 메인 카드 박스 */}
+      {/* 메인 카드 */}
       <Box
         sx={{
           border: "1px solid #e0e0e0",
@@ -53,7 +67,7 @@ export default function WorkcenterPlanView() {
         }}
       >
         <Typography variant="h6" gutterBottom>
-          작업장 별 생산계획
+          📌 작업장 별 생산 계획 (총 {total}건)
         </Typography>
 
         <Box sx={{ flex: 1, overflow: "auto" }}>
@@ -74,7 +88,7 @@ export default function WorkcenterPlanView() {
                 fontWeight: "bold",
               },
               border: 0,
-              minWidth: "900px",
+              minWidth: "1200px",
             }}
           />
         </Box>
